requirements.txt
streamlit 
langchain
langchain_community
langchain_groq
youtube-transcript-api
python-dotenv
validators
ipython
unstructured